tests/wpt/tests/intersection-observer/svg-intersection-with-fractional-bounds.html

<!DOCTYPE html>
<title>IntersectionObserver observing an SVG &lt;rect> with fractional bounds element</title>
<meta name="viewport" content="width=device-width,initial-scale=1">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/intersection-observer-test-utils.js"></script>
<style>
.spacer {
  height: calc(100vh + 100px);
}
</style>
<div class="spacer"></div>
<svg viewBox="0 0 200 100" width="4" display="block">
  <rect id="target" x="50" y="5" width="100" height="50" fill="blue"/>
</svg>
<div class="spacer"></div>
<script>
const viewportWidth = document.documentElement.clientWidth;
const viewportHeight = document.documentElement.clientHeight;
setup(() => {
  window.entries = [];
  window.target = document.getElementById("target");
  window.targetRect = target.getBoundingClientRect();
});
var epsilon = 0.0001;
runTestCycle(function() {
  assert_true(!!target, "target exists");
  const observer = new IntersectionObserver(function(changes) {
    entries = entries.concat(changes);
  });
  observer.observe(target);
  entries = entries.concat(observer.takeRecords());
  assert_equals(entries.length, 0, "No initial notifications");
  runTestCycle(step0, "Initial Observation");
});
function step0() {
  document.scrollingElement.scrollTop = 300;
  runTestCycle(step1, "Intersection observation after scrolling 300");
  // The numbers in brackets are target client rect; intersection rect;
  // and root bounds.
  checkLastEntry(entries, 0, [
    // the 8 pixels comes from the html body padding.
    1 + 8, 1 + 8 + targetRect.width, 8 + viewportHeight + 100 + 0.1, viewportHeight + 100 + 0.1 + 8 + targetRect.height,
    0, 0, 0, 0,
    0, viewportWidth, 0, viewportHeight,
    false,
  ], epsilon);
}
function step1() {
  document.scrollingElement.scrollTop = 100;
  runTestCycle(step2, "Intersection observation after scrolling 100");
  checkLastEntry(entries, 1, [
    1 + 8, 1 + 8 + targetRect.width, 8 + viewportHeight - 200 + 0.1, viewportHeight - 200 + 0.1 + 8 + targetRect.height,
    1 + 8, 1 + 8 + targetRect.width, 8 + viewportHeight - 200 + 0.1, viewportHeight - 200 + 0.1 + 8 + targetRect.height,
    0, viewportWidth, 0, viewportHeight,
    true,
  ], epsilon);
}
function step2() {
  document.scrollingElement.scrollTop = 0;
  checkLastEntry(entries, 2, [
    1 + 8, 1 + 8 + targetRect.width, 8 + viewportHeight + 0.1, viewportHeight + 0.1 + 8 + targetRect.height,
    0, 0, 0, 0,
    0, viewportWidth, 0, viewportHeight,
    false,
  ], epsilon);
}
</script>
