LayoutTests/imported/w3c/web-platform-tests/intersection-observer/svg-intersection-with-fractional-bounds-expected.txt


PASS IntersectionObserver observing an SVG <rect> with fractional bounds element
FAIL Initial Observation assert_approx_equals: entries[0].boundingClientRect.left expected 9 +/- 0.0001 but got 8
FAIL Intersection observation after scrolling 300 assert_approx_equals: entries[1].boundingClientRect.left expected 9 +/- 0.0001 but got 8
FAIL Intersection observation after scrolling 100 assert_approx_equals: entries[2].boundingClientRect.left expected 9 +/- 0.0001 but got 8
